move auth to separate file & create follow routing
Check is the requesting user is following already. Check is the user is private and append the requesting user to the requested users pending collection.  If the user is public append the requesting user to the requested user's followers collection and append the requested user to the requesting user's following collection.